Minor fix to count in inventory items. Fix EnemyScaling to allow for decimals

+++ BrotatoServer/Models/JSON/EnemyScaling.cs
@@ -5,11 +5,11 @@ namespace BrotatoServer.Models.JSON;
 public class EnemyScaling
 {
     [JsonProperty("health")]
-    public int Health { get; set; }
+    public decimal Health { get; set; }
 
     [JsonProperty("damage")]
-    public int Damage { get; set; }
+    public decimal Damage { get; set; }
 
     [JsonProperty("speed")]
-    public int Speed { get; set; }
+    public decimal Speed { get; set; }
 }
